chore(reorganized-structure-directories): sectioning directories
