+++ src/App.js
@@ -1,6 +1,6 @@
 import axios from "axios";
-import { BookInfo } from "./components/info/book";
-import { DataSourceRenderPattern } from "./components/data-source-with-render-pattern";
+import { BookInfo } from "./section2-3-layout-and-container-components/info/book";
+import { DataSourceRenderPattern } from "./section2-3-layout-and-container-components/data-source-with-render-pattern";
 
 const getData = async (url) => {
   const res = await axios.get(url);
