+++ src/section2-3-layout-and-container-components/authors/SmallListItems.jsx
