+++ src/section2-3-layout-and-container-components/books/SmallListItems.jsx
