+++ src/section2-3-layout-and-container-components/data-source-with-render-pattern.jsx
