+++ src/section2-3-layout-and-container-components/index.js
@@ -1,0 +1,9 @@
+export * from './info';
+export * from './books';
+export * from './users';
+export * from './authors';
+export { Modal } from './Modal';
+export { DataSource } from './data-source'
+export { SplitScreen } from './split-screen';
+export { ResourceLoader } from './resource-loader';
+export { DataSourceRenderPattern } from './data-source-with-render-pattern';
