+++ src/section2-3-layout-and-container-components/info/index.js
