+++ src/section2-3-layout-and-container-components/split-screen.jsx
